# specs/001-quickflow-productivity/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: QuickFlow API
  version: 1.0.0
  description: >
    REST API of QuickFlow (single-user personal productivity). Source of truth for the backend springdoc swagger
    and the generated typescript-angular client. Errors use RFC 9457 application/problem+json.
servers:
- {url: 'http://localhost:8080'}
...
paths:
...
  /api/habits:
    get:
      tags: [Habits]
      operationId: listHabits
      parameters:
      - name: active
        in: query
        required: false
        schema: {type: boolean}
        description: Omit for all habits
      responses:
        '200':
          description: Habits with stats, newest first
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Habit'}
        '400': {$ref: '#/components/responses/BadRequest'}
    post:
      tags: [Habits]
      operationId: createHabit
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/HabitRequest'}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Habit'}
        '400': {$ref: '#/components/responses/BadRequest'}
  /api/habits/{id}:
    parameters:
    - name: id
      in: path
      required: true
      schema: {type: integer, format: int64}
    get:
      tags: [Habits]
      operationId: getHabit
      responses:
        '200':
          description: The habit
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Habit'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
    put:
      tags: [Habits]
      operationId: updateHabit
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/HabitRequest'}
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Habit'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
    delete:
      tags: [Habits]
      operationId: deleteHabit
      responses:
        '204': {description: Deleted with its completions}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/habits/{id}/deactivate:
    parameters:
    - name: id
      in: path
      required: true
      schema: {type: integer, format: int64}
    post:
      tags: [Habits]
      operationId: deactivateHabit
      responses:
        '200':
          description: Deactivated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Habit'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/habits/{id}/activate:
    parameters:
    - name: id
      in: path
      required: true
      schema: {type: integer, format: int64}
    post:
      tags: [Habits]
      operationId: activateHabit
      responses:
        '200':
          description: Activated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Habit'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/habits/{id}/completions:
    parameters:
    - name: id
      in: path
      required: true
      schema: {type: integer, format: int64}
    get:
      tags: [Habits]
      operationId: listHabitCompletions
      responses:
        '200':
          description: Completions, newest first
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/HabitCompletion'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
    post:
      tags: [Habits]
      operationId: completeHabit
      summary: Record a completion (default today); one per habit per date
      requestBody:
        required: false
        content:
          application/json:
            schema: {$ref: '#/components/schemas/HabitCompletionRequest'}
      responses:
        '201':
          description: Recorded; returns the habit with updated stats
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Habit'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/habits/{id}/completions/{date}:
    parameters:
    - name: id
      in: path
      required: true
      schema: {type: integer, format: int64}
    - name: date
      in: path
      required: true
      schema: {type: string, format: date}
    delete:
      tags: [Habits]
      operationId: uncompleteHabit
      summary: Remove the completion for a date
      responses:
        '200':
          description: Removed; returns the habit with updated stats
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Habit'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
...
components:
  responses:
    BadRequest:
      description: Validation error
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    Conflict:
      description: Conflict (e.g. duplicate habit completion)
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem details
      properties:
        type: {type: string}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        instance: {type: string}
        errors:
          type: array
          items: {$ref: '#/components/schemas/FieldError'}
    FieldError:
      type: object
      required: [field, message]
      properties:
        field: {type: string}
        message: {type: string}
...
    HabitFrequency:
      type: string
      enum: [DAILY, WEEKLY]
    HabitRequest:
      type: object
      required: [name, frequency]
      properties:
        name: {type: string, minLength: 1, maxLength: 150}
        description:
          type: [string, 'null']
          maxLength: 2000
        frequency: {$ref: '#/components/schemas/HabitFrequency'}
    Habit:
      type: object
      required: [id, name, frequency, createdAt, active, completedToday, doneForCurrentPeriod, currentStreak, completionRate]
      properties:
        id: {type: integer, format: int64}
        name: {type: string}
        description:
          type: [string, 'null']
        frequency: {$ref: '#/components/schemas/HabitFrequency'}
        createdAt: {type: string, format: date-time}
        active: {type: boolean}
        completedToday: {type: boolean}
        doneForCurrentPeriod: {type: boolean}
        currentStreak: {type: integer}
        completionRate: {type: integer, minimum: 0, maximum: 100}
        lastCompletedDate:
          type: [string, 'null']
          format: date
    HabitCompletionRequest:
      type: object
      properties:
        date:
          type: [string, 'null']
          format: date
          description: Defaults to today; must not be in the future
    HabitCompletion:
      type: object
      required: [id, habitId, completionDate, createdAt]
      properties:
        id: {type: integer, format: int64}
        habitId: {type: integer, format: int64}
        completionDate: {type: string, format: date}
        createdAt: {type: string, format: date-time}
...